Adds logging, and an install script
Adds logging to the script so that it can display the date and
timestamp.

Also adds an install script to copy the appropriate files to the
appropriate locations.

It can be started/controlled by 'sudo service mythpywii restart', and
should start on bootup.

Also note that it restarts bluetooth. a bit distructive, but that's the
best solution to me.